refactor: 필터링 된 스튜디오 조회 기능 이전 (Studio => ConceptStudio) - Endpoint 수정 - DTO 통일 (StudioResponse로 통일)
