+++ src/main/java/com/toucheese/conceptstudio/service/ConceptStudioService.java
@@ -1,44 +1,47 @@
 package com.toucheese.conceptstudio.service;
 
-import com.toucheese.conceptstudio.dto.ConceptStudioResponse;
 import com.toucheese.conceptstudio.entity.ConceptStudio;
 import com.toucheese.conceptstudio.repository.ConceptStudioRepository;
-import com.toucheese.studio.entity.Studio;
-import com.toucheese.studio.repository.StudioRepository;
-
+import com.toucheese.studio.dto.StudioResponse;
+import com.toucheese.studio.entity.Location;
+import com.toucheese.studio.repository.StudioRepositoryImpl;
 import lombok.RequiredArgsConstructor;
-
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class ConceptStudioService {
 
 	private final ConceptStudioRepository conceptStudioRepository;
-	private final StudioRepository studioRepository;
+	private final StudioRepositoryImpl studioRepositoryImpl;
 
-	public Page<ConceptStudioResponse> getStudiosByConceptId(Long conceptId, int page) {
-		final int PAGE_SIZE = 10;
+	private static final int PAGE_SIZE = 10;
 
+	public Page<StudioResponse> getStudiosByConceptId(Long conceptId, int page) {
 		Pageable pageable = PageRequest.of(page, PAGE_SIZE);
+		return conceptStudioRepository.findByConceptId(conceptId, pageable).map( conceptStudio ->
+				StudioResponse.of(conceptStudio.getStudio())
+		);
+	}
 
-		Page<ConceptStudio> conceptStudios = conceptStudioRepository.findByConceptId(conceptId, pageable);
-
-		return conceptStudios.map(cs -> {
-			Studio studio = cs.getStudio();
-
-			List<String> imageUrls = studio.getImages().stream()
-				.map(image -> image.getUrl())
-				.collect(Collectors.toList());
-
-			return ConceptStudioResponse.of(studio, imageUrls);
-		});
+	/**
+	 * 필터링 된 스튜디오 목록을 이름으로 정렬하여 조회한다.
+	 * @param page 현재 페이지
+	 * @param conceptId 선택된 컨셉 요소 아이디
+	 * @param price 가격순 필터링 요소
+	 * @param rating 인기순 필터링 요소
+	 * @param locations 지역 필터링 요소
+	 * @return 현재 페이지에 해당하는 필터링 후 정렬된 스튜디오 목록
+	 */
+	public Page<StudioResponse> getFilteredStudiosOrderByName(int page, Long conceptId, Integer price, Float rating, List<Location> locations) {
+		Pageable pageable = PageRequest.of(page, PAGE_SIZE);
+		return studioRepositoryImpl.getFilteredStudiosOrderByName(price, rating, locations, conceptId, pageable)
+				.map(StudioResponse::of);
 	}
 
 }
